Move guts of Track() class back from moofimage.py to wozimage.py. The Track() class bit and nibble processing methods are useful in both, and since MoofTrack is a subclass of Track, it seems worthwile to keep the methods in the base class.
